fix: address upstream review findings

+++ ports/muda/src/platform_impl/gtk/mod.rs
@@ -145,8 +145,8 @@ fn strip_mnemonic(text: &str) -> String {
 
 pub struct Menu {
     id: MenuId,
-    instances: HashMap<u32, GtkMenuBar>,
-    ctx_menu_id: u32,
+    instances: HashMap<usize, GtkMenuBar>,
+    ctx_menu_id: usize,
     children: Vec<Rc<RefCell<MenuChild>>>,
 }
 
@@ -155,7 +155,7 @@ impl Menu {
         Self {
             id: id.unwrap_or_else(|| MenuId(COUNTER.next().to_string())),
             instances: HashMap::new(),
-            ctx_menu_id: COUNTER.next(),
+            ctx_menu_id: COUNTER.next() as usize,
             children: Vec::new(),
         }
     }
@@ -191,7 +191,7 @@ impl Menu {
         Ok(())
     }
 
-    pub fn add_menu_item_with_id(&mut self, item: &dyn IsMenuItem, id: u32) -> crate::Result<()> {
+    pub fn add_menu_item_with_id(&mut self, item: &dyn IsMenuItem, id: usize) -> crate::Result<()> {
         for (menu_id, menu_bar) in self.instances.iter().filter(|m| *m.0 == id) {
             let parent_menu = menu_bar.menu();
             let gtk_item =
@@ -272,7 +272,7 @@ impl Menu {
         W: gtk::prelude::IsA<gtk::Widget>,
         C: gtk::prelude::IsA<gtk::Widget>,
     {
-        let id = window.as_ptr() as u32;
+        let id = window.as_ptr() as usize;
 
         let Some(app) = window.application() else {
             return Err(crate::Error::GtkWindowWithoutApplication);
@@ -326,7 +326,7 @@ impl Menu {
         W: gtk::prelude::IsA<gtk::Window>,
         W: gtk::prelude::IsA<gtk::Widget>,
     {
-        let id = window.as_ptr() as u32;
+        let id = window.as_ptr() as usize;
 
         let Some(menu_bar) = self.instances.remove(&id) else {
             return Err(crate::Error::NotInitialized);
@@ -349,7 +349,7 @@ impl Menu {
     where
         W: gtk::prelude::IsA<gtk::Window>,
     {
-        let id = window.as_ptr() as u32;
+        let id = window.as_ptr() as usize;
         let Some(menu_bar) = self.instances.get(&id) else {
             return Err(crate::Error::NotInitialized);
         };
@@ -361,7 +361,7 @@ impl Menu {
     where
         W: gtk::prelude::IsA<gtk::Window>,
     {
-        let id = window.as_ptr() as u32;
+        let id = window.as_ptr() as usize;
         let Some(menu_bar) = self.instances.get(&id) else {
             return Err(crate::Error::NotInitialized);
         };
@@ -374,7 +374,7 @@ impl Menu {
     where
         W: gtk::prelude::IsA<gtk::Window>,
     {
-        let id = window.as_ptr() as u32;
+        let id = window.as_ptr() as usize;
         self.instances
             .get(&id)
             .map(|m| m.menu_bar().is_visible())
@@ -385,7 +385,7 @@ impl Menu {
     where
         W: gtk::prelude::IsA<gtk::Window>,
     {
-        let id = window.as_ptr() as u32;
+        let id = window.as_ptr() as usize;
         self.instances.get(&id).map(|m| m.menu_bar().clone())
     }
 
@@ -469,14 +469,14 @@ enum GtkMenuChild {
         parent_menu: gio::Menu,
     },
     Submenu {
-        id: u32,
+        id: usize,
         item: gio::MenuItem,
         menu: gio::Menu,
         app: gtk::Application,
         parent_menu: gio::Menu,
     },
     ContextMenu {
-        id: u32,
+        id: usize,
         widget: gtk::PopoverMenu,
         menu: gio::Menu,
         app: gtk::Application,
@@ -484,7 +484,7 @@ enum GtkMenuChild {
 }
 
 impl GtkMenuChild {
-    fn id(&self) -> u32 {
+    fn id(&self) -> usize {
         match self {
             GtkMenuChild::Submenu { id, .. } => *id,
             GtkMenuChild::ContextMenu { id, .. } => *id,
@@ -544,8 +544,8 @@ pub struct MenuChild {
 
     type_: MenuItemType,
 
-    instances: HashMap<u32, Vec<GtkMenuChild>>,
-    ctx_menu_id: u32,
+    instances: HashMap<usize, Vec<GtkMenuChild>>,
+    ctx_menu_id: usize,
     children: Vec<Rc<RefCell<MenuChild>>>,
 
     action: Option<gio::SimpleAction>,
@@ -635,7 +635,7 @@ impl MenuChild {
             key_accelerator: None,
             predefined_item_type: None,
             type_: MenuItemType::Submenu,
-            ctx_menu_id: COUNTER.next(),
+            ctx_menu_id: COUNTER.next() as usize,
             instances: HashMap::new(),
             children: Vec::new(),
             action: None,
@@ -647,7 +647,7 @@ impl MenuChild {
     fn create_gtk_item_for_submenu(
         &mut self,
         app: &gtk::Application,
-        menu_id: u32,
+        menu_id: usize,
         parent_menu: &gio::Menu,
     ) -> crate::Result<gio::MenuItem> {
         let menu = gio::Menu::new();
@@ -665,7 +665,7 @@ impl MenuChild {
             self.action = Some(action);
         }
 
-        let id = COUNTER.next();
+        let id = COUNTER.next() as usize;
         let child = GtkMenuChild::Submenu {
             item: item.clone(),
             menu,
@@ -713,7 +713,7 @@ impl MenuChild {
         Ok(())
     }
 
-    pub fn add_menu_item_with_id(&self, item: &dyn IsMenuItem, id: u32) -> crate::Result<()> {
+    pub fn add_menu_item_with_id(&self, item: &dyn IsMenuItem, id: usize) -> crate::Result<()> {
         for menus in self.instances.values() {
             for gtk_child in menus.iter().filter(|m| m.id() == id) {
                 let parent_menu = gtk_child.menu();
@@ -907,7 +907,7 @@ impl MenuChild {
     fn create_gtk_item_for_menu_item(
         &mut self,
         app: &gtk::Application,
-        menu_id: u32,
+        menu_id: usize,
         parent_menu: &gio::Menu,
     ) -> crate::Result<gio::MenuItem> {
         let detailed_action = self.detailed_action();
@@ -1102,7 +1102,7 @@ impl MenuChild {
     fn create_gtk_item_for_check_menu_item(
         &mut self,
         app: &gtk::Application,
-        menu_id: u32,
+        menu_id: usize,
         parent_menu: &gio::Menu,
     ) -> crate::Result<gio::MenuItem> {
         let detailed_action = self.detailed_action();
@@ -1209,7 +1209,7 @@ impl MenuChild {
     fn create_gtk_item_for_icon_menu_item(
         &mut self,
         app: &gtk::Application,
-        menu_id: u32,
+        menu_id: usize,
         parent_menu: &gio::Menu,
     ) -> crate::Result<gio::MenuItem> {
         let detailed_action = self.detailed_action();
@@ -1259,7 +1259,7 @@ impl MenuChild {
     fn create_gtk_item_for_predefined_menu_item(
         &mut self,
         app: &gtk::Application,
-        menu_id: u32,
+        menu_id: usize,
         parent_menu: &gio::Menu,
     ) -> crate::Result<gio::MenuItem> {
         let predefined_item_type = self.predefined_item_type.clone().unwrap();
@@ -1494,7 +1494,7 @@ impl dyn IsMenuItem + '_ {
     fn make_gtk_menu_item(
         &self,
         app: &gtk::Application,
-        menu_id: u32,
+        menu_id: usize,
         parent_menu: &gio::Menu,
     ) -> crate::Result<gio::MenuItem> {
         let kind = self.kind();
